asdf

+++ 2학기 과제/11주차 과제/문제4_바이트단위 출력.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 
 
-long asdf(FILE* fp) {
+int asdf(FILE* fp) {
 	int fpos = ftell(fp);
 	return fpos;
 }
